Make powermenu lockscreen visibility configurable [1/2]
This gives the user the option to limit how their power menu looks on
the lockscreen while enabling all options once the device is unlocked.

--Example--
Lockscreen - Power and Reboot
Homescreen - Power, Reboot, Advanced Reboot, Screenshot and Airplane

In the past this was either all or nothing. Having a full power meny
on the lockscreen just looks off considering that some options are only
essential once the device is unlocked.

Change-Id: I153b79c3d6dc0be9af048b160a7ac056a364d3d2
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>