Ensure screen pinning toast is themed
This forces this toast to change with our themes. Previously this toast
would not change with overlays. Even setting it to a certain color and
call it with an overlay, NOTHING. So what we're doing here is checking
for the dark theme overlay and changing the color accordingly.

- Move our ThemeAccentsUtils class to the core/java/com/android/internal/statusbar
so we can call it from not only the SystemUI but other packages. This will
reduce the amount of code in packages like DU-Tweaks, Updater and Settings.

- Consolidate the code with our themes tile

Change-Id: Ia77d53802b73ac1157f3cc8ce5887a3e50555382
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>