Ensure that locked recents apps are respected
Introduced in the commit below is the ability to kill any app from
recents in a safe way. The problem with this is that it kills any app
at any time no matter what. This shouldn't be the case. If an app is
locked then is safe from recents 'clear-all' and should be safe from
this method of killing apps.

https://github.com/DirtyUnicorns/android_frameworks_base/commit/060fe20e0ef7c28605f4854d3895a9da597e5e89

Change-Id: Ia84ef3fcf607dc3924da91fbfa4607db3407cc72
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>