Hide freeform task header icon
So the question is, why hide this icon?

The answer is simple. This 'mod' is not supported by Google. Is not even a
mod that's on by default. You literally have to use ADB to make it appear
and even then is just asking for problems because not all apps are optimized
to support it (which is the reason even Google doesn't have it on by default).

To learn more about how to enable it via adb, check the link below
https://medium.com/sentio-superbook/how-to-enable-multi-window-in-android-o-678cced03db2

Change-Id: I6bcc517db2ee9f2f0531114f36ea7bded2a18d56
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>